rename: change a bottomBar file path
