+++ app/src/main/java/com/jukco/waitforme/ui/components/BottomNaviBar.kt
@@ -1,7 +1,10 @@
-package com.jukco.waitforme.ui.navi
+package com.jukco.waitforme.ui.components
 
+import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.Icon
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
@@ -9,6 +12,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavDestination
@@ -17,14 +21,8 @@ import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-
-val BOTTOM_NAVI_ROUTE = listOf<BottomNaviItem>(
-    BottomNaviItem.PopsManagement,
-    BottomNaviItem.PopsList,
-    BottomNaviItem.Bookmark,
-    BottomNaviItem.WaitInfo,
-    BottomNaviItem.MyInfo,
-)
+import com.jukco.waitforme.R
+import com.jukco.waitforme.ui.navi.Route
 
 @Composable
 fun BottomNaviBar(navController: NavHostController) {
@@ -87,3 +85,23 @@ private fun BottomNaviBarPreview() {
         currentDestination = null,
     )
 }
+
+sealed class BottomNaviItem(
+    @StringRes val title: Int,
+    val icon: ImageVector,
+    val route: String,
+) {
+    object PopsManagement : BottomNaviItem(R.string.navi_pops_management, Icons.Filled.Favorite, Route.PopsManagement.name)
+    object PopsList : BottomNaviItem(R.string.navi_pops_list, Icons.Filled.Favorite, Route.PopsList.name)
+    object Bookmark : BottomNaviItem(R.string.navi_bookmark, Icons.Filled.Favorite, Route.Bookmark.name)
+    object WaitInfo : BottomNaviItem(R.string.navi_wait_info, Icons.Filled.Favorite, Route.WaitInfo.name)
+    object MyInfo : BottomNaviItem(R.string.navi_my_info, Icons.Filled.Favorite, Route.MyInfo.name)
+}
+
+val BOTTOM_NAVI_ROUTE = listOf<BottomNaviItem>(
+    BottomNaviItem.PopsManagement,
+    BottomNaviItem.PopsList,
+    BottomNaviItem.Bookmark,
+    BottomNaviItem.WaitInfo,
+    BottomNaviItem.MyInfo,
+)
